+++ dev.sh
@@ -13,5 +13,6 @@ npx nodemon \
   -w 'test-of-complex-query.js' \
   -w 'test-of-dataset-proxy.js' \
   -w 'test-of-server.js' \
+  -w 'test-of-controller-language.js' \
   -e 'js' \
   -x 'bash test-all.sh'