+++ test-of-controller-language.sh
@@ -1,0 +1,3 @@
+npx peggy --format globals --export-var ControllerLanguage -o controller-language.js controller-language.pegjs
+node fix.js
+node test-of-controller-language.js